[extension-manager] * add script service security * add extensions to upgrade in the install confirmation UI
git-svn-id: https://svn.xwiki.org/svnroot/xwiki/contrib/sandbox@31188 f329d543-caf0-0310-9063-dda96c69346f

+++ xwiki-extension/xwiki-extension-manager/src/main/java/org/xwiki/extension/internal/script/ExtensionManagerScriptService.java
@@ -2,6 +2,7 @@ package org.xwiki.extension.internal.script;
 
 import java.util.List;
 
+import org.xwiki.bridge.DocumentAccessBridge;
 import org.xwiki.component.annotation.Component;
 import org.xwiki.component.annotation.Requirement;
 import org.xwiki.extension.Extension;
@@ -11,6 +12,7 @@ import org.xwiki.extension.InstallException;
 import org.xwiki.extension.LocalExtension;
 import org.xwiki.extension.ResolveException;
 import org.xwiki.extension.UninstallException;
+import org.xwiki.extension.internal.VersionManager;
 import org.xwiki.extension.repository.LocalExtensionRepository;
 import org.xwiki.script.service.ScriptService;
 
@@ -21,11 +23,19 @@ public class ExtensionManagerScriptService implements ScriptService
     private ExtensionManager extensionManager;
 
     @Requirement
+    private DocumentAccessBridge documentAccessBridge;
+
+    @Requirement
+    private VersionManager versionManager;
+
+    @Requirement
     private LocalExtensionRepository localExtensionRepository;
 
     public LocalExtension install(String id, String version) throws InstallException
     {
-        // TODO: check rights
+        if (!this.documentAccessBridge.hasProgrammingRights()) {
+            return null;
+        }
 
         return this.extensionManager.installExtension(new ExtensionId(id, version));
     }
@@ -37,7 +47,9 @@ public class ExtensionManagerScriptService implements ScriptService
 
     public void uninstall(String id) throws UninstallException
     {
-        // TODO: check rights
+        if (!this.documentAccessBridge.hasProgrammingRights()) {
+            return;
+        }
 
         this.extensionManager.uninstallExtension(id);
     }
@@ -50,5 +62,15 @@ public class ExtensionManagerScriptService implements ScriptService
     public List<LocalExtension> getInstalledExtensions()
     {
         return this.localExtensionRepository.getLocalExtensions();
+    }
+    
+    public LocalExtension getInstalledExtension(String id)
+    {
+        return this.localExtensionRepository.getLocalExtension(id);
+    }
+
+    public VersionManager getVersionManager()
+    {
+        return this.versionManager;
     }
 }
